requirements.txt
appdirs==1.4.4
asgiref==3.2.3
awscli==1.18.49
boto3==1.13.24
botocore==1.16.24
camelcase==0.2
certifi==2020.6.20
cffi==1.14.0
click==6.7
colorama==0.4.3
configobj==5.0.6
cryptography==2.9.2
DateTime==4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-heroku==0.3.1
django-messages==0.6.0
django-session-security==2.6.6
django-shell-plus==1.1.7
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
jmespath==0.9.5
json-syntax==2.3.1
jsons==1.1.2
Markdown==3.2.1
Pillow==7.1.1
pipenv==2020.6.2
prompt-toolkit==1.0.18
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.7
pycryptodomex==3.9.4
Pygments==2.6.1
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
saws==0.4.3
six==1.14.0
sqlparse==0.3.1
typish==1.3.2
urllib3==1.25.9
virtualenv==20.0.23
virtualenv-clone==0.5.4
wcwidth==0.1.9
whitenoise==5.0.1
zope.interface==4.7.1
